# openapi.yml
openapi: 3.0.3
info:
  title: Luga Store API
  version: 1.0.0
  description: API documentation for the Luga Store backend, listing all customer, shop manager/owner, and administrator endpoints.
security:
  - CookieAuth: []
paths:
...
  /auth/login:
    post:
      tags: [Authentication]
      summary: Login to an account and receive session cookie
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful, sets session cookie.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/register:
    post:
      tags: [Authentication]
      summary: Register a new customer account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Registration successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /auth/forgot-password:
    post:
      tags: [Authentication]
      summary: Request password reset link
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
      responses:
        '200':
          description: Reset email sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /auth/reset-password:
    post:
      tags: [Authentication]
      summary: Reset account password using token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Password reset successful
        '400':
          $ref: '#/components/responses/BadRequest'

  # --- INVITATIONS ---
...
  /categories:
    get:
      tags: [Categories (Global)]
      summary: Get all global or shop-linked categories
      security: []
      parameters:
        - name: shopId
          in: query
          schema:
            type: integer
        - name: pageNumber
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 10
        - name: sortBy
          in: query
          schema:
            type: string
            default: name
        - name: sortOrder
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: asc
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Categories retrieved

    post:
      tags: [Categories (Global)]
      summary: Create a new category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertCategoryRequest'
      responses:
        '200':
          description: Category created successfully
  /categories/{id}:
    put:
      tags: [Categories (Global)]
      summary: Update a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertCategoryRequest'
      responses:
        '200':
          description: Category updated successfully

    delete:
      tags: [Categories (Global)]
      summary: Delete a category (high performance)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: shopId
          in: query
          schema:
            type: integer
      responses:
      '200':
        description: Category deleted successfully
        content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  type: string
  /categories/{id}/status:
    patch:
      tags: [Categories (Global)]
      summary: Toggle category active status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCategoryActiveStatusRequest'
      responses:
        '200':
          description: Active status updated successfully
  /categories/reorder:
    post:
      tags: [Categories (Global)]
      summary: Reorder categories
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReorderCategoriesRequest'
      responses:
        '200':
          description: Reordered successfully

  # --- SHOP-LEVEL CATEGORIES ---
...
  /shops/{shopId}/categories/{id}:
    put:
      tags: [Categories (Shop)]
      summary: Update a category under a shop
      parameters:
        - name: shopId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShopUpsertCategoryRequest'
      responses:
        '200':
          description: Category updated successfully

    delete:
      tags: [Categories (Shop)]
      summary: Delete a shop category
      parameters:
        - name: shopId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
      '200':
        description: Category deleted successfully
        content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  type: string
  /shops/{shopId}/categories/{id}/status:
    patch:
      tags: [Categories (Shop)]
      summary: Toggle active status of a shop category
      parameters:
        - name: shopId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCategoryActiveStatusRequest'
      responses:
        '200':
          description: Status toggled successfully
  /shops/{shopId}/categories/reorder:
    post:
      tags: [Categories (Shop)]
      summary: Reorder categories of a shop
      parameters:
        - name: shopId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReorderCategoriesRequest'
      responses:
        '200':
          description: Reordered successfully

  # --- GLOBAL PRODUCTS ---
...
  /profile/cart:
    get:
      tags: [Cart]
      summary: Retrieve the current user's shopping cart
      responses:
        '200':
          description: Cart details

    post:
      tags: [Cart]
      summary: Add an item to the cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddToCartRequest'
      responses:
        '200':
          description: Added successfully

    delete:
      tags: [Cart]
      summary: Clear all items in the cart
      responses:
      '200':
        description: Cart cleared successfully
        content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  type: string
  /profile/cart/{id}:
    put:
      tags: [Cart]
      summary: Update quantity or size of a cart item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCartRequest'
      responses:
        '200':
          description: Updated successfully

    delete:
      tags: [Cart]
      summary: Remove an item from the cart
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
      '200':
        description: Item removed successfully
        content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  type: string

  # --- WISHLIST ---
...
  /admin/orders:
    get:
      tags: [Orders (Admin)]
      summary: Get customer orders list
      parameters:
        - name: customerId
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: List of orders
  /admin/orders/{orderId}/status:
    patch:
      tags: [Orders (Admin)]
      summary: Update status of an order
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
        - name: customerId
          in: query
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: integer
              description: OrderStatus enum value (e.g. 0=Pending, 1=Processing, 2=Shipped, 3=Delivered, 4=Cancelled)
      responses:
        '200':
          description: Order status updated successfully

  # --- MEDIA MANAGEMENT ---
...
  /customer/newsletter/subscribe:
    post:
      tags: [Newsletter]
      summary: Subscribe to the newsletter
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewsletterRequest'
      responses:
        '200':
          description: Subscription successful
  /customer/newsletter/validate-unsubscribe/{token}:
    get:
      tags: [Newsletter]
      summary: Validate an unsubscribe token
      security: []
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Token is valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  email:
                    type: string
        '404':
          description: Invalid unsubscribe link
  /customer/newsletter/confirm-unsubscribe:
    post:
      tags: [Newsletter]
      summary: Confirm unsubscription using token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UnsubscribeConfirmationRequest'
      responses:
        '200':
          description: Unsubscribed successfully
  /customer/newsletter/unsubscribe:
    post:
      tags: [Newsletter]
      summary: Direct unsubscribe request
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewsletterRequest'
      responses:
        '200':
          description: Unsubscribed successfully

  # --- USERS (Admin-only management) ---
...
  /users/customers/{id}:
    delete:
      tags: [Users (Admin)]
      summary: Delete a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
      '200':
        description: Customer deleted successfully
        content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  type: string
...
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: .AspNetCore.Identity.Application
  responses:
    BadRequest:
      description: Invalid request parameters or validation failed
    Unauthorized:
      description: Unauthenticated or missing permissions
...
  schemas:
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        rememberMe:
          type: boolean
          default: false
    RegisterRequest:
      type: object
      required: [email, password, firstName, lastName]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
    ForgotPasswordRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
    ResetPasswordRequest:
      type: object
      required: [email, token, newPassword]
      properties:
        email:
          type: string
          format: email
        token:
          type: string
        newPassword:
          type: string
...
    UpsertCategoryRequest:
      type: object
      required: [name]
      properties:
        shopId:
          type: integer
          nullable: true
        name:
          type: string
        description:
          type: string
    ShopUpsertCategoryRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
    UpdateCategoryActiveStatusRequest:
      type: object
      required: [isActive]
      properties:
        isActive:
          type: boolean
        shopId:
          type: integer
          nullable: true
    ReorderCategoriesRequest:
      type: object
      required: [orders]
      properties:
        shopId:
          type: integer
          nullable: true
        orders:
          type: array
          items:
            type: object
            required: [id, displayOrder]
            properties:
              id:
                type: integer
              displayOrder:
                type: integer
...
    AddToCartRequest:
      type: object
      required: [productId, size, quantity]
      properties:
        productId:
          type: integer
        size:
          type: string
          enum: [XS, S, M, L, XL, XXL, XXXL]
        quantity:
          type: integer
    UpdateCartRequest:
      type: object
      required: [size, quantity]
      properties:
        size:
          type: string
          enum: [XS, S, M, L, XL, XXL, XXXL]
        quantity:
          type: integer
...
    NewsletterRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
    UnsubscribeConfirmationRequest:
      type: object
      required: [token]
      properties:
        token:
          type: string
...